+++ mode/core/src/test/java/org/apache/shardingsphere/mode/metadata/persist/service/config/RepositoryTuplePersistServiceTest.java
@@ -17,6 +17,7 @@
 
 package org.apache.shardingsphere.mode.metadata.persist.service.config;
 
+import org.apache.shardingsphere.mode.metadata.persist.config.RepositoryTuplePersistService;
 import org.apache.shardingsphere.mode.spi.repository.PersistRepository;
 import org.apache.shardingsphere.mode.node.tuple.RepositoryTuple;
 import org.junit.jupiter.api.BeforeEach;