+++ mode/core/src/test/java/org/apache/shardingsphere/mode/metadata/persist/service/metadata/table/TableRowDataPersistServiceTest.java
@@ -18,8 +18,9 @@
 package org.apache.shardingsphere.mode.metadata.persist.service.metadata.table;
 
 import org.apache.shardingsphere.infra.metadata.database.schema.model.ShardingSphereTable;
-import org.apache.shardingsphere.infra.metadata.statistics.ShardingSphereTableData;
-import org.apache.shardingsphere.infra.yaml.data.pojo.YamlShardingSphereRowData;
+import org.apache.shardingsphere.infra.metadata.statistics.TableStatistics;
+import org.apache.shardingsphere.infra.yaml.data.pojo.YamlRowStatistics;
+import org.apache.shardingsphere.mode.metadata.persist.metadata.table.TableRowDataPersistService;
 import org.apache.shardingsphere.mode.spi.repository.PersistRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -57,7 +58,7 @@ class TableRowDataPersistServiceTest {
     
     @Test
     void assertPersistWithRows() {
-        YamlShardingSphereRowData rowData = new YamlShardingSphereRowData();
+        YamlRowStatistics rowData = new YamlRowStatistics();
         rowData.setUniqueKey("foo_key");
         persistService.persist("foo_db", "foo_schema", "foo_tbl", Collections.singletonList(rowData));
         verify(repository).persist("/statistics/databases/foo_db/schemas/foo_schema/tables/foo_tbl/foo_key", "uniqueKey: foo_key" + System.lineSeparator());
@@ -65,7 +66,7 @@ class TableRowDataPersistServiceTest {
     
     @Test
     void assertDelete() {
-        YamlShardingSphereRowData rowData = new YamlShardingSphereRowData();
+        YamlRowStatistics rowData = new YamlRowStatistics();
         rowData.setUniqueKey("foo_key");
         persistService.delete("foo_db", "foo_schema", "foo_tbl", Collections.singletonList(rowData));
         verify(repository).delete("/statistics/databases/foo_db/schemas/foo_schema/tables/foo_tbl/foo_key");
@@ -77,7 +78,7 @@ class TableRowDataPersistServiceTest {
         when(repository.query("/statistics/databases/foo_db/schemas/foo_schema/tables/foo_tbl/foo_tbl")).thenReturn("uniqueKey: foo_key" + System.lineSeparator());
         ShardingSphereTable table = mock(ShardingSphereTable.class);
         when(table.getName()).thenReturn("foo_tbl");
-        ShardingSphereTableData actual = persistService.load("foo_db", "foo_schema", table);
+        TableStatistics actual = persistService.load("foo_db", "foo_schema", table);
         assertThat(actual.getName(), is("foo_tbl"));
         assertThat(actual.getRows().size(), is(1));
         assertThat(actual.getRows().iterator().next().getUniqueKey(), is("foo_key"));
@@ -89,7 +90,7 @@ class TableRowDataPersistServiceTest {
         when(repository.query("/statistics/databases/foo_db/schemas/foo_schema/tables/foo_tbl/foo_tbl")).thenReturn("");
         ShardingSphereTable table = mock(ShardingSphereTable.class);
         when(table.getName()).thenReturn("foo_tbl");
-        ShardingSphereTableData actual = persistService.load("foo_db", "foo_schema", table);
+        TableStatistics actual = persistService.load("foo_db", "foo_schema", table);
         assertThat(actual.getName(), is("foo_tbl"));
         assertTrue(actual.getRows().isEmpty());
     }